refactor test names
