+++ app/src/androidTest/java/com/group4sweng/scranplan/Templates/ExampleInstrumentedTest.java
@@ -1,10 +1,12 @@
-package com.group4sweng.scranplan;
+package com.group4sweng.scranplan.templates;
 
 import android.content.Context;
 
 import androidx.test.platform.app.InstrumentationRegistry;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.rule.ActivityTestRule;
+
+import com.group4sweng.scranplan.Login;
 
 import org.junit.Rule;
 import org.junit.Test;
