+++ app/src/test/java/com/group4sweng/scranplan/SoundHandler/AudioUrlTest.java
@@ -14,7 +14,7 @@ import static org.junit.Assert.assertTrue;
  * Written 22/3/20 NBillis
  * (c) CoDev 2020
  */
-public class AudioURLTest {
+public class AudioUrlTest {
 
     @Test
     public void testPlayingURLSounds() throws Exception{
